backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.9.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
sentence-transformers>=2.7.0
pinecone-client>=2.2.4
groq>=0.4.1
python-docx==1.1.0
PyPDF2==3.0.1
python-dotenv==1.0.0
aiofiles==23.2.1